Fix primary cell identification
We ran into an issue that all available cells are not of type
"NoneConnection" (available, but not connected), but of type
"SecondaryConnection" which made the being identified as
connected.

Fix this by checking for "PrimaryConnection" type instead
of checking for "not NoneConnection".

A similar issue war reported to netmonster-core before:

https://github.com/mroczis/netmonster-core/issues/53